stage-2/requirements.txt
sentence-transformers
numpy<2.0
faiss-gpu